Important changes
I have decided to remove the first parameter of each function (which was basically the information about the pattern of workbench). This information is now obtained within the function itself. I have not done this update for all functions yet, so this needs to be fixed at some point.

I have also added a new function to set the structure, to convert from metric to label, and from border to ROI.